% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3-package.R
\docType{package}
\name{gh3-package}
\alias{gh3}
\alias{gh3-package}
\title{gh3: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramnathv/gh3}
  \item Report bugs at \url{https://github.com/ramnathv/gh3/issues}
}

}
\author{
\strong{Maintainer}: Ramnath Vaidyanathan \email{ramnath@datacamp.com}

}
\keyword{internal}
