% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_emojis_get}
\alias{gh_emojis_get}
\title{Lists all the emojis available to use on GitHub.}
\usage{
gh_emojis_get(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Lists all the emojis available to use on GitHub.
}
