% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gists_delete_star}
\alias{gh_gists_delete_star}
\title{Unstar a gist.}
\usage{
gh_gists_delete_star(id, ...)
}
\arguments{
\item{id}{Id of gist.}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Unstar a gist.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
