% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gists_get_by_id}
\alias{gh_gists_get_by_id}
\title{Get a single gist.}
\usage{
gh_gists_get_by_id(id, ...)
}
\arguments{
\item{id}{Id of gist.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get a single gist.
}
