% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gists_get_comments}
\alias{gh_gists_get_comments}
\title{List comments on a gist.}
\usage{
gh_gists_get_comments(id, ...)
}
\arguments{
\item{id}{Id of gist.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List comments on a gist.
}
