% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gists_get_public}
\alias{gh_gists_get_public}
\title{List all public gists.}
\usage{
gh_gists_get_public(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List all public gists.
}
