% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gists_get_star}
\alias{gh_gists_get_star}
\title{Check if a gist is starred.}
\usage{
gh_gists_get_star(id, ...)
}
\arguments{
\item{id}{Id of gist.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Check if a gist is starred.
}
