% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gists_patch_by_id}
\alias{gh_gists_patch_by_id}
\title{Edit a gist.}
\usage{
gh_gists_patch_by_id(id, body, ...)
}
\arguments{
\item{id}{Id of gist.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Edit a gist.
}
