% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gists_patch_comments_by_comment_id}
\alias{gh_gists_patch_comments_by_comment_id}
\title{Edit a comment.}
\usage{
gh_gists_patch_comments_by_comment_id(id, commentId, body, ...)
}
\arguments{
\item{id}{Id of gist.}

\item{commentId}{Id of comment.}

\item{body}{body}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Edit a comment.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
