% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gists_post}
\alias{gh_gists_post}
\title{Create a gist.}
\usage{
gh_gists_post(body, ...)
}
\arguments{
\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Create a gist.
}
