% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gists_post_forks}
\alias{gh_gists_post_forks}
\title{Fork a gist.}
\usage{
gh_gists_post_forks(id, ...)
}
\arguments{
\item{id}{Id of gist.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Fork a gist.
}
