% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gists_put_star}
\alias{gh_gists_put_star}
\title{Star a gist.}
\usage{
gh_gists_put_star(id, ...)
}
\arguments{
\item{id}{Id of gist.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Star a gist.
}
