% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gitignore_get_templates}
\alias{gh_gitignore_get_templates}
\title{Listing available templates.}
\usage{
gh_gitignore_get_templates(...)
}
\arguments{
\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
List all templates available to pass as an option when creating a repository.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
