% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gitignore_get_templates_by_language}
\alias{gh_gitignore_get_templates_by_language}
\title{Get a single template.}
\usage{
gh_gitignore_get_templates_by_language(language, ...)
}
\arguments{
\item{language}{language}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Get a single template.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
