% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_legacy_get_issues_search_by_keyword}
\alias{gh_legacy_get_issues_search_by_keyword}
\title{Find issues by state and keyword.}
\usage{
gh_legacy_get_issues_search_by_keyword(keyword, state, owner, repository,
  ...)
}
\arguments{
\item{keyword}{The search term.}

\item{state}{Indicates the state of the issues to return. Can be either open or
closed.}

\item{owner}{owner}

\item{repository}{repository}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Find issues by state and keyword.
}
