% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_legacy_get_repos_search_by_keyword}
\alias{gh_legacy_get_repos_search_by_keyword}
\title{Find repositories by keyword. Note, this legacy method does not follow the v3 pagination pattern. This method returns up to 100 results per page and pages can be fetched using the start_page parameter.}
\usage{
gh_legacy_get_repos_search_by_keyword(keyword, ...)
}
\arguments{
\item{keyword}{The search term}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Find repositories by keyword. Note, this legacy method does not follow the v3 pagination pattern. This method returns up to 100 results per page and pages can be fetched using the start_page parameter.
}
