% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_legacy_get_user_email_by_email}
\alias{gh_legacy_get_user_email_by_email}
\title{This API call is added for compatibility reasons only.}
\usage{
gh_legacy_get_user_email_by_email(email, ...)
}
\arguments{
\item{email}{The email address}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
This API call is added for compatibility reasons only.
}
