% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_legacy_get_user_search_by_keyword}
\alias{gh_legacy_get_user_search_by_keyword}
\title{Find users by keyword.}
\usage{
gh_legacy_get_user_search_by_keyword(keyword, ...)
}
\arguments{
\item{keyword}{The search term}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Find users by keyword.
}
