% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_markdown_post}
\alias{gh_markdown_post}
\title{Render an arbitrary Markdown document}
\usage{
gh_markdown_post(body, ...)
}
\arguments{
\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Render an arbitrary Markdown document
}
