% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_markdown_post_raw}
\alias{gh_markdown_post_raw}
\title{Render a Markdown document in raw mode}
\usage{
gh_markdown_post_raw(...)
}
\arguments{
\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Render a Markdown document in raw mode
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
