% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_meta_get}
\alias{gh_meta_get}
\title{This gives some information about GitHub.com, the service.}
\usage{
gh_meta_get(...)
}
\arguments{
\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
This gives some information about GitHub.com, the service.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
