% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_networks_get_events}
\alias{gh_networks_get_events}
\title{List public events for a network of repositories.}
\usage{
gh_networks_get_events(owner, repo, ...)
}
\arguments{
\item{owner}{Name of the owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List public events for a network of repositories.
}
