% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_notifications_get_threads_subscription}
\alias{gh_notifications_get_threads_subscription}
\title{Get a Thread Subscription.}
\usage{
gh_notifications_get_threads_subscription(id, ...)
}
\arguments{
\item{id}{Id of thread.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get a Thread Subscription.
}
