% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_notifications_put}
\alias{gh_notifications_put}
\title{Mark as read.}
\usage{
gh_notifications_put(body, ...)
}
\arguments{
\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Marking a notification as "read" removes it from the default view on
GitHub.com.
}
