% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_notifications_put_threads_subscription}
\alias{gh_notifications_put_threads_subscription}
\title{Set a Thread Subscription.}
\usage{
gh_notifications_put_threads_subscription(id, body, ...)
}
\arguments{
\item{id}{Id of thread.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
This lets you subscribe to a thread, or ignore it. Subscribing to a thread is
unnecessary if the user is already subscribed to the repository. Ignoring a
thread will mute all future notifications (until you comment or get
}
