% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_orgs_delete_members_by_username}
\alias{gh_orgs_delete_members_by_username}
\title{Remove a member.}
\usage{
gh_orgs_delete_members_by_username(org, username, ...)
}
\arguments{
\item{org}{Name of organisation.}

\item{username}{Name of the user.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Removing a user from this list will remove them from all teams and they will no
longer have any access to the organization's repositories.
}
