% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_orgs_get_events}
\alias{gh_orgs_get_events}
\title{List public events for an organization.}
\usage{
gh_orgs_get_events(org, ...)
}
\arguments{
\item{org}{Name of organisation.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List public events for an organization.
}
