% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_orgs_get_public_members}
\alias{gh_orgs_get_public_members}
\title{Public members list.}
\usage{
gh_orgs_get_public_members(org, ...)
}
\arguments{
\item{org}{Name of organisation.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Members of an organization can choose to have their membership publicized or
not.
}
