% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_orgs_get_repos}
\alias{gh_orgs_get_repos}
\title{List repositories for the specified org.}
\usage{
gh_orgs_get_repos(org, ...)
}
\arguments{
\item{org}{Name of organisation.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List repositories for the specified org.
}
