% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_orgs_put_public_members_by_username}
\alias{gh_orgs_put_public_members_by_username}
\title{Publicize a user's membership.}
\usage{
gh_orgs_put_public_members_by_username(org, username, ...)
}
\arguments{
\item{org}{Name of organisation.}

\item{username}{Name of the user.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Publicize a user's membership.
}
