% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_rate_limit_get}
\alias{gh_rate_limit_get}
\title{Get your current rate limit status}
\usage{
gh_rate_limit_get(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Note: Accessing this endpoint does not count against your rate limit.
}
