% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_delete_downloads_by_download_id}
\alias{gh_repos_delete_downloads_by_download_id}
\title{Deprecated. Delete a download.}
\usage{
gh_repos_delete_downloads_by_download_id(owner, repo, downloadId, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{downloadId}{Id of download.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Deprecated. Delete a download.
}
