% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_delete_issues_comments_by_comment_id}
\alias{gh_repos_delete_issues_comments_by_comment_id}
\title{Delete a comment.}
\usage{
gh_repos_delete_issues_comments_by_comment_id(owner, repo, commentId, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{commentId}{ID of comment.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Delete a comment.
}
