% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_delete_issues_labels_by_name}
\alias{gh_repos_delete_issues_labels_by_name}
\title{Remove a label from an issue.}
\usage{
gh_repos_delete_issues_labels_by_name(owner, repo, number, name, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{number}{Number of issue.}

\item{name}{Name of the label.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Remove a label from an issue.
}
