% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_delete_milestones_by_number}
\alias{gh_repos_delete_milestones_by_number}
\title{Delete a milestone.}
\usage{
gh_repos_delete_milestones_by_number(owner, repo, number, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{number}{Number of milestone.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Delete a milestone.
}
