% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_assignees}
\alias{gh_repos_get_assignees}
\title{List assignees.}
\usage{
gh_repos_get_assignees(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
This call lists all the available assignees (owner + collaborators) to which
issues may be assigned.
}
