% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_by_repo}
\alias{gh_repos_get_by_repo}
\title{Get repository.}
\usage{
gh_repos_get_by_repo(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get repository.
}
