% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_comments_by_comment_id}
\alias{gh_repos_get_comments_by_comment_id}
\title{Get a single commit comment.}
\usage{
gh_repos_get_comments_by_comment_id(owner, repo, commentId, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{commentId}{Id of comment.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get a single commit comment.
}
