% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_commits}
\alias{gh_repos_get_commits}
\title{List commits on a repository.}
\usage{
gh_repos_get_commits(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List commits on a repository.
}
