% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_commits_by_sha_code}
\alias{gh_repos_get_commits_by_sha_code}
\title{Get a single commit.}
\usage{
gh_repos_get_commits_by_sha_code(owner, repo, shaCode, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{shaCode}{SHA-1 code of the commit.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get a single commit.
}
