% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_commits_comments}
\alias{gh_repos_get_commits_comments}
\title{List comments for a single commitList comments for a single commit.}
\usage{
gh_repos_get_commits_comments(owner, repo, shaCode, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{shaCode}{SHA-1 code of the commit.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List comments for a single commitList comments for a single commit.
}
