% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_commits_status}
\alias{gh_repos_get_commits_status}
\title{Get the combined Status for a specific Ref}
\usage{
gh_repos_get_commits_status(owner, repo, ref, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{ref}{ref}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
The Combined status endpoint is currently available for developers to preview.
During the preview period, the API may change without advance notice. Please
see the blog post for full details. To access this endpoint during the preview
period, you must provide a custom media type in the Accept header:
application/vnd.github.she-hulk-preview+json
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
