% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_contributors}
\alias{gh_repos_get_contributors}
\title{Get list of contributors.}
\usage{
gh_repos_get_contributors(owner, repo, anon, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{anon}{Set to 1 or true to include anonymous contributors in results.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get list of contributors.
}
