% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_deployments}
\alias{gh_repos_get_deployments}
\title{Users with pull access can view deployments for a repository}
\usage{
gh_repos_get_deployments(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Users with pull access can view deployments for a repository
}
