% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_events}
\alias{gh_repos_get_events}
\title{Get list of repository events.}
\usage{
gh_repos_get_events(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get list of repository events.
}
