% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_git_commits_by_sha_code}
\alias{gh_repos_get_git_commits_by_sha_code}
\title{Get a Commit.}
\usage{
gh_repos_get_git_commits_by_sha_code(owner, repo, shaCode, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{shaCode}{SHA-1 code.}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Get a Commit.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
