% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_git_trees_by_sha_code}
\alias{gh_repos_get_git_trees_by_sha_code}
\title{Get a Tree.}
\usage{
gh_repos_get_git_trees_by_sha_code(owner, repo, shaCode, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{shaCode}{Tree SHA.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get a Tree.
}
