% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_hooks_by_hook_id}
\alias{gh_repos_get_hooks_by_hook_id}
\title{Get single hook.}
\usage{
gh_repos_get_hooks_by_hook_id(owner, repo, hookId, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{hookId}{Id of hook.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get single hook.
}
