% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_issues}
\alias{gh_repos_get_issues}
\title{List issues for a repository.}
\usage{
gh_repos_get_issues(owner, repo, filter, state, labels, sort, direction,
  ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{filter}{Issues assigned to you / created by you / mentioning you / you're
subscribed to updates for / All issues the authenticated user can see}

\item{state}{state}

\item{labels}{String list of comma separated Label names. Example - bug,ui,@high.}

\item{sort}{sort}

\item{direction}{direction}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
List issues for a repository.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
