% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_issues_by_number}
\alias{gh_repos_get_issues_by_number}
\title{Get a single issue}
\usage{
gh_repos_get_issues_by_number(owner, repo, number, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{number}{Number of issue.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get a single issue
}
