% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_issues_comments}
\alias{gh_repos_get_issues_comments}
\title{List comments in a repository.}
\usage{
gh_repos_get_issues_comments(owner, repo, number, ...)

gh_repos_get_issues_comments(owner, repo, number, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{number}{Number of issue.}

\item{...}{Additional parameters to pass to gh::gh.}

\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List comments in a repository.

List comments on an issue.
}
