% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_issues_events}
\alias{gh_repos_get_issues_events}
\title{List issue events for a repository.}
\usage{
gh_repos_get_issues_events(owner, repo, number, ...)

gh_repos_get_issues_events(owner, repo, number, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{number}{Number of issue.}

\item{...}{Additional parameters to pass to gh::gh.}

\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List issue events for a repository.

List events for an issue.
}
