% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_labels}
\alias{gh_repos_get_labels}
\title{List all labels for this repository.}
\usage{
gh_repos_get_labels(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List all labels for this repository.
}
