% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_milestones_labels}
\alias{gh_repos_get_milestones_labels}
\title{Get labels for every issue in a milestone.}
\usage{
gh_repos_get_milestones_labels(owner, repo, number, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{number}{Number of milestone.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get labels for every issue in a milestone.
}
