% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_notifications}
\alias{gh_repos_get_notifications}
\title{List your notifications in a repository}
\usage{
gh_repos_get_notifications(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List all notifications for the current user.
}
