% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_pulls_comments}
\alias{gh_repos_get_pulls_comments}
\title{List comments in a repository.}
\usage{
gh_repos_get_pulls_comments(owner, repo, number, ...)

gh_repos_get_pulls_comments(owner, repo, number, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{number}{Id of pull.}

\item{...}{Additional parameters to pass to gh::gh.}

\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
By default, Review Comments are ordered by ascending ID.
}
