% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_readme}
\alias{gh_repos_get_readme}
\title{Get the README.}
\usage{
gh_repos_get_readme(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
This method returns the preferred README for a repository.
}
