% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_stats_commit_activity}
\alias{gh_repos_get_stats_commit_activity}
\title{Get the last year of commit activity data.}
\usage{
gh_repos_get_stats_commit_activity(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Returns the last year of commit activity grouped by week. The days array is a
group of commits per day, starting on Sunday.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
