% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_stats_contributors}
\alias{gh_repos_get_stats_contributors}
\title{Get contributors list with additions, deletions, and commit counts.}
\usage{
gh_repos_get_stats_contributors(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get contributors list with additions, deletions, and commit counts.
}
