% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_stats_participation}
\alias{gh_repos_get_stats_participation}
\title{Get the weekly commit count for the repo owner and everyone else.}
\usage{
gh_repos_get_stats_participation(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get the weekly commit count for the repo owner and everyone else.
}
