% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_stats_punch_card}
\alias{gh_repos_get_stats_punch_card}
\title{Get the number of commits per hour in each day.}
\usage{
gh_repos_get_stats_punch_card(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Each array contains the day number, hour number, and number of commits 0-6
Sunday - Saturday 0-23 Hour of day Number of commits
}
\details{
For example, \link{2, 14, 25} indicates that there were 25 total commits, during the
2.00pm hour on Tuesdays. All times are based on the time zone of individual
commits.
}
