% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_statuses_by_ref}
\alias{gh_repos_get_statuses_by_ref}
\title{List Statuses for a specific Ref.}
\usage{
gh_repos_get_statuses_by_ref(owner, repo, ref, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{ref}{Ref to list the statuses from. It can be a SHA, a branch name, or a tag
name.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List Statuses for a specific Ref.
}
