% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_patch_by_repo}
\alias{gh_repos_patch_by_repo}
\title{Edit repository.}
\usage{
gh_repos_patch_by_repo(owner, repo, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Edit repository.
}
