% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_patch_labels_by_name}
\alias{gh_repos_patch_labels_by_name}
\title{Update a label.}
\usage{
gh_repos_patch_labels_by_name(owner, repo, name, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{name}{Name of the label.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Update a label.
}
