% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_patch_pulls_comments_by_comment_id}
\alias{gh_repos_patch_pulls_comments_by_comment_id}
\title{Edit a comment.}
\usage{
gh_repos_patch_pulls_comments_by_comment_id(owner, repo, commentId, body,
  ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{commentId}{Id of comment.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Edit a comment.
}
