% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_post_commits_comments}
\alias{gh_repos_post_commits_comments}
\title{Create a commit comment.}
\usage{
gh_repos_post_commits_comments(owner, repo, shaCode, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{shaCode}{SHA-1 code of the commit.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Create a commit comment.
}
