% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_post_deployments_statuses}
\alias{gh_repos_post_deployments_statuses}
\title{Create a Deployment Status}
\usage{
gh_repos_post_deployments_statuses(owner, repo, id, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{id}{The Deployment ID to list the statuses from.}

\item{body}{body}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Users with push access can create deployment statuses for a given deployment:
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
