% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_post_forks}
\alias{gh_repos_post_forks}
\title{Create a fork.}
\usage{
gh_repos_post_forks(owner, repo, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Forking a Repository happens asynchronously. Therefore, you may have to wai a
short period before accessing the git objects. If this takes longer than 5
minutes, be sure to contact Support.
}
