% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_post_git_tags}
\alias{gh_repos_post_git_tags}
\title{Create a Tag Object.}
\usage{
gh_repos_post_git_tags(owner, repo, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Note that creating a tag object does not create the reference that makes a tag
in Git. If you want to create an annotated tag in Git, you have to do this call
to create the tag object, and then create the refs/tags/\link{tag} reference. If you
want to create a lightweight tag, you only have to create the tag reference -
this call would be unnecessary.
}
