% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_post_hooks_tests}
\alias{gh_repos_post_hooks_tests}
\title{Test a push hook.}
\usage{
gh_repos_post_hooks_tests(owner, repo, hookId, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{hookId}{Id of hook.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
This will trigger the hook with the latest push to the current repository if
the hook is subscribed to push events. If the hook is not subscribed to push
events, the server will respond with 204 but no test POST will be generated.
Note: Previously /repos/:owner/:repo/hooks/:id/tes
}
