% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_post_releases}
\alias{gh_repos_post_releases}
\title{Create a release}
\usage{
gh_repos_post_releases(owner, repo, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Users with push access to the repository can create a release.
}
