% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_post_statuses_by_ref}
\alias{gh_repos_post_statuses_by_ref}
\title{Create a Status.}
\usage{
gh_repos_post_statuses_by_ref(owner, repo, ref, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{ref}{Ref to list the statuses from. It can be a SHA, a branch name, or a tag
name.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Create a Status.
}
