% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_put_collaborators_by_user}
\alias{gh_repos_put_collaborators_by_user}
\title{Add collaborator.}
\usage{
gh_repos_put_collaborators_by_user(owner, repo, user, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{user}{Login of the user.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Add collaborator.
}
