% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_put_contents_by_path}
\alias{gh_repos_put_contents_by_path}
\title{Create a file.}
\usage{
gh_repos_put_contents_by_path(owner, repo, path, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{path}{path}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Create a file.
}
