% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repositories_get}
\alias{gh_repositories_get}
\title{List all public repositories.}
\usage{
gh_repositories_get(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
This provides a dump of every public repository, in the order that they were
created. Note: Pagination is powered exclusively by the since parameter. is the
Link header to get the URL for the next page of repositories.
}
