% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_search_get_code}
\alias{gh_search_get_code}
\title{Search code.}
\usage{
gh_search_get_code(q, ...)
}
\arguments{
\item{q}{The search terms. This can be any combination of the supported code search
parameters: 'Search In' Qualifies which fields are searched. With this
qualifier you can restrict the search to just the file contents, the file path,
or both. 'Languages' Searches code based on the language it's written in.
'Forks' Filters repositories based on the number of forks, and/or whether code
from forked repositories should be included in the results at all. 'Size' Finds
files that match a certain size (in bytes). 'Path' Specifies the path that the
resulting file must be at. 'Extension' Matches files with a certain extension.
'Users' or 'Repositories' Limits searches to a specific user or repository.}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Search code.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
