% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_search_get_repositories}
\alias{gh_search_get_repositories}
\title{Search repositories.}
\usage{
gh_search_get_repositories(q, ...)
}
\arguments{
\item{q}{The search terms. This can be any combination of the supported repository
search parameters: 'Search In' Qualifies which fields are searched. With this
qualifier you can restrict the search to just the repository name, description,
readme, or any combination of these. 'Size' Finds repositories that match a
certain size (in kilobytes). 'Forks' Filters repositories based on the number
of forks, and/or whether forked repositories should be included in the results
at all. 'Created' and 'Last Updated' Filters repositories based on times of
creation, or when they were last updated. 'Users or Repositories' Limits
searches to a specific user or repository. 'Languages' Searches repositories
based on the language they are written in. 'Stars' Searches repositories based
on the number of stars.}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Search repositories.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
