% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_search_get_users}
\alias{gh_search_get_users}
\title{Search users.}
\usage{
gh_search_get_users(q, ...)
}
\arguments{
\item{q}{The search terms. This can be any combination of the supported user search
parameters: 'Search In' Qualifies which fields are searched. With this
qualifier you can restrict the search to just the username, public email, full
name, location, or any combination of these. 'Repository count' Filters users
based on the number of repositories they have. 'Location' Filter users by the
location indicated in their profile. 'Language' Search for users that have
repositories that match a certain language. 'Created' Filter users based on
when they joined. 'Followers' Filter users based on the number of followers
they have.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Search users.
}
