% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_teams_delete_members_by_username}
\alias{gh_teams_delete_members_by_username}
\title{The "Remove team member" API is deprecated and is scheduled for removal in the next major version of the API. We recommend using the Remove team membership API instead. It allows you to remove both active and pending memberships.}
\usage{
gh_teams_delete_members_by_username(teamId, username, ...)
}
\arguments{
\item{teamId}{Id of team.}

\item{username}{Name of a member.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Remove team member. In order to remove a user from a team, the authenticated
user must have 'admin' permissions to the team or be an owner of the org that
the team is associated with. NOTE This does not delete the user, it just remove
them from the team.
}
