% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_teams_delete_memberships_by_username}
\alias{gh_teams_delete_memberships_by_username}
\title{Remove team membership.}
\usage{
gh_teams_delete_memberships_by_username(teamId, username, ...)
}
\arguments{
\item{teamId}{Id of team.}

\item{username}{Name of a member.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
In order to remove a membership between a user and a team, the authenticated
user must have 'admin' permissions to the team or be an owner of the
organization that the team is associated with. NOTE: This does not delete the
user, it just removes their membership from the team.
}
