% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_teams_delete_repos_by_repo}
\alias{gh_teams_delete_repos_by_repo}
\title{In order to remove a repository from a team, the authenticated user must be an owner of the org that the team is associated with. NOTE: This does not delete the repository, it just removes it from the team.}
\usage{
gh_teams_delete_repos_by_repo(teamId, owner, repo, ...)
}
\arguments{
\item{teamId}{Id of team.}

\item{owner}{Name of a repository owner.}

\item{repo}{Name of a repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
In order to remove a repository from a team, the authenticated user must be an owner of the org that the team is associated with. NOTE: This does not delete the repository, it just removes it from the team.
}
