% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_teams_get_by_team_id}
\alias{gh_teams_get_by_team_id}
\title{Get team.}
\usage{
gh_teams_get_by_team_id(teamId, ...)
}
\arguments{
\item{teamId}{Id of team.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get team.
}
