% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_teams_get_members}
\alias{gh_teams_get_members}
\title{List team members.}
\usage{
gh_teams_get_members(teamId, ...)
}
\arguments{
\item{teamId}{Id of team.}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
In order to list members in a team, the authenticated user must be a member of
the team.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
