% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_teams_get_members_by_username}
\alias{gh_teams_get_members_by_username}
\title{The "Get team member" API is deprecated and is scheduled for removal in the next major version of the API. We recommend using the Get team membership API instead. It allows you to get both active and pending memberships.}
\usage{
gh_teams_get_members_by_username(teamId, username, ...)
}
\arguments{
\item{teamId}{Id of team.}

\item{username}{Name of a member.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get team member. In order to get if a user is a member of a team, the
authenticated user mus be a member of the team.
}
