% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_teams_get_repos}
\alias{gh_teams_get_repos}
\title{List team repos}
\usage{
gh_teams_get_repos(teamId, ...)
}
\arguments{
\item{teamId}{Id of team.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List team repos
}
