% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_teams_get_repos_by_repo}
\alias{gh_teams_get_repos_by_repo}
\title{Check if a team manages a repository}
\usage{
gh_teams_get_repos_by_repo(teamId, owner, repo, ...)
}
\arguments{
\item{teamId}{Id of team.}

\item{owner}{Name of a repository owner.}

\item{repo}{Name of a repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Check if a team manages a repository
}
