% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_teams_patch_by_team_id}
\alias{gh_teams_patch_by_team_id}
\title{Edit team.}
\usage{
gh_teams_patch_by_team_id(teamId, body, ...)
}
\arguments{
\item{teamId}{Id of team.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
In order to edit a team, the authenticated user must be an owner of the org
that the team is associated with.
}
