% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_teams_put_members_by_username}
\alias{gh_teams_put_members_by_username}
\title{The API (described below) is deprecated and is scheduled for removal in the next major version of the API. We recommend using the Add team membership API instead. It allows you to invite new organization members to your teams.}
\usage{
gh_teams_put_members_by_username(teamId, username, ...)
}
\arguments{
\item{teamId}{Id of team.}

\item{username}{Name of a member.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Add team member. In order to add a user to a team, the authenticated user must
have 'admin' permissions to the team or be an owner of the org that the team is
associated with.
}
