% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_teams_put_repos_by_repo}
\alias{gh_teams_put_repos_by_repo}
\title{In order to add a repository to a team, the authenticated user must be an owner of the org that the team is associated with. Also, the repository must be owned by the organization, or a direct fork of a repository owned by the organization.}
\usage{
gh_teams_put_repos_by_repo(teamId, owner, repo, ...)
}
\arguments{
\item{teamId}{Id of team.}

\item{owner}{Name of a organization.}

\item{repo}{Name of a repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
In order to add a repository to a team, the authenticated user must be an owner of the org that the team is associated with. Also, the repository must be owned by the organization, or a direct fork of a repository owned by the organization.
}
