% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_delete_emails}
\alias{gh_user_delete_emails}
\title{Delete email address(es).}
\usage{
gh_user_delete_emails(body, ...)
}
\arguments{
\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
You can include a single email address or an array of addresses.
}
