% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_delete_following_by_username}
\alias{gh_user_delete_following_by_username}
\title{Unfollow a user.}
\usage{
gh_user_delete_following_by_username(username, ...)
}
\arguments{
\item{username}{Name of user.}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Unfollowing a user requires the user to be logged in and authenticated with
basic auth or OAuth with the user:follow scope.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
