% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_delete_starred_by_repo}
\alias{gh_user_delete_starred_by_repo}
\title{Unstar a repository}
\usage{
gh_user_delete_starred_by_repo(owner, repo, ...)
}
\arguments{
\item{owner}{Name of a repository owner.}

\item{repo}{Name of a repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Unstar a repository
}
