% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_get_emails}
\alias{gh_user_get_emails}
\title{List email addresses for a user.}
\usage{
gh_user_get_emails(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
In the final version of the API, this method will return an array of hashes
with extended information for each email address indicating if the address has
been verified and if it's primary email address for GitHub. Until API v3 is
finalized, use the application/vnd.github.v3 media type to get other response
format.
}
