% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_get_followers}
\alias{gh_user_get_followers}
\title{List the authenticated user's followers}
\usage{
gh_user_get_followers(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List the authenticated user's followers
}
