% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_get_orgs}
\alias{gh_user_get_orgs}
\title{List public and private organizations for the authenticated user.}
\usage{
gh_user_get_orgs(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List public and private organizations for the authenticated user.
}
