% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_get_repos}
\alias{gh_user_get_repos}
\title{List repositories for the authenticated user. Note that this does not include}
\usage{
gh_user_get_repos(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
repositories owned by organizations which the user can access. You can lis user
organizations and list organization repositories separately.
}
