% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_get_subscriptions_by_repo}
\alias{gh_user_get_subscriptions_by_repo}
\title{Check if you are watching a repository.}
\usage{
gh_user_get_subscriptions_by_repo(owner, repo, ...)
}
\arguments{
\item{owner}{Name of the owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Check if you are watching a repository.
}
