% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_get_teams}
\alias{gh_user_get_teams}
\title{List all of the teams across all of the organizations to which the authenticated user belongs. This method requires user or repo scope when authenticating via OAuth.}
\usage{
gh_user_get_teams(...)
}
\arguments{
\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
List all of the teams across all of the organizations to which the authenticated user belongs. This method requires user or repo scope when authenticating via OAuth.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
