% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_patch}
\alias{gh_user_patch}
\title{Update the authenticated user.}
\usage{
gh_user_patch(body, ...)
}
\arguments{
\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Update the authenticated user.
}
