% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_post_keys}
\alias{gh_user_post_keys}
\title{Create a public key.}
\usage{
gh_user_post_keys(body, ...)
}
\arguments{
\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Create a public key.
}
