% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_post_repos}
\alias{gh_user_post_repos}
\title{Create a new repository for the authenticated user. OAuth users must supply}
\usage{
gh_user_post_repos(body, ...)
}
\arguments{
\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
repo scope.
}
