% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_put_following_by_username}
\alias{gh_user_put_following_by_username}
\title{Follow a user.}
\usage{
gh_user_put_following_by_username(username, ...)
}
\arguments{
\item{username}{Name of user.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Following a user requires the user to be logged in and authenticated with basic
auth or OAuth with the user:follow scope.
}
