% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_users_get_by_username}
\alias{gh_users_get_by_username}
\title{Get a single user.}
\usage{
gh_users_get_by_username(username, ...)
}
\arguments{
\item{username}{Name of user.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get a single user.
}
