% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_users_get_events_orgs_by_org}
\alias{gh_users_get_events_orgs_by_org}
\title{This is the user's organization dashboard. You must be authenticated as the user to view this.}
\usage{
gh_users_get_events_orgs_by_org(username, org, ...)
}
\arguments{
\item{username}{Name of user.}

\item{org}{org}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
This is the user's organization dashboard. You must be authenticated as the user to view this.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
