% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_users_get_followers}
\alias{gh_users_get_followers}
\title{List a user's followers}
\usage{
gh_users_get_followers(username, ...)
}
\arguments{
\item{username}{Name of user.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List a user's followers
}
