% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_users_get_following_by_target_user}
\alias{gh_users_get_following_by_target_user}
\title{Check if one user follows another.}
\usage{
gh_users_get_following_by_target_user(username, targetUser, ...)
}
\arguments{
\item{username}{Name of user.}

\item{targetUser}{Name of user.}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Check if one user follows another.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
