% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_users_get_keys}
\alias{gh_users_get_keys}
\title{List public keys for a user.}
\usage{
gh_users_get_keys(username, ...)
}
\arguments{
\item{username}{Name of user.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Lists the verified public keys for a user. This is accessible by anyone.
}
