% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_users_get_orgs}
\alias{gh_users_get_orgs}
\title{List all public organizations for a user.}
\usage{
gh_users_get_orgs(username, ...)
}
\arguments{
\item{username}{Name of user.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List all public organizations for a user.
}
