% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_users_get_received_events}
\alias{gh_users_get_received_events}
\title{These are events that you'll only see public events.}
\usage{
gh_users_get_received_events(username, ...)
}
\arguments{
\item{username}{Name of user.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
These are events that you'll only see public events.
}
