% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_users_get_received_events_public}
\alias{gh_users_get_received_events_public}
\title{List public events that a user has received}
\usage{
gh_users_get_received_events_public(username, ...)
}
\arguments{
\item{username}{Name of user.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List public events that a user has received
}
