% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_users_get_repos}
\alias{gh_users_get_repos}
\title{List public repositories for the specified user.}
\usage{
gh_users_get_repos(username, ...)
}
\arguments{
\item{username}{Name of user.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List public repositories for the specified user.
}
