% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_users_get_starred}
\alias{gh_users_get_starred}
\title{List repositories being starred by a user.}
\usage{
gh_users_get_starred(username, ...)
}
\arguments{
\item{username}{Name of user.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List repositories being starred by a user.
}
