% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_feeds_get}
\alias{gh_feeds_get}
\title{List Feeds.}
\usage{
gh_feeds_get(...)
}
\arguments{
\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
GitHub provides several timeline resources in Atom format. The Feeds API lists
all the feeds available to the authenticating user.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
