% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gists_delete_star}
\alias{gh_gists_delete_star}
\title{Unstar a gist.}
\usage{
gh_gists_delete_star(id, ...)
}
\arguments{
\item{id}{Id of gist.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Unstar a gist.
}
