% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gists_get}
\alias{gh_gists_get}
\title{List the authenticated user's gists or if called anonymously, this will}
\usage{
gh_gists_get(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
return all public gists.
}
