% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gists_get_comments_by_comment_id}
\alias{gh_gists_get_comments_by_comment_id}
\title{Get a single comment.}
\usage{
gh_gists_get_comments_by_comment_id(id, commentId, ...)
}
\arguments{
\item{id}{Id of gist.}

\item{commentId}{Id of comment.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get a single comment.
}
