% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gists_get_starred}
\alias{gh_gists_get_starred}
\title{List the authenticated user's starred gists.}
\usage{
gh_gists_get_starred(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List the authenticated user's starred gists.
}
