% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_gitignore_get_templates_by_language}
\alias{gh_gitignore_get_templates_by_language}
\title{Get a single template.}
\usage{
gh_gitignore_get_templates_by_language(language, ...)
}
\arguments{
\item{language}{language}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get a single template.
}
