% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_legacy_get_user_email_by_email}
\alias{gh_legacy_get_user_email_by_email}
\title{This API call is added for compatibility reasons only.}
\usage{
gh_legacy_get_user_email_by_email(email, ...)
}
\arguments{
\item{email}{The email address}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
This API call is added for compatibility reasons only.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
