% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_markdown_post}
\alias{gh_markdown_post}
\title{Render an arbitrary Markdown document}
\usage{
gh_markdown_post(body, ...)
}
\arguments{
\item{body}{body}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Render an arbitrary Markdown document
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
