% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_markdown_post_raw}
\alias{gh_markdown_post_raw}
\title{Render a Markdown document in raw mode}
\usage{
gh_markdown_post_raw(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Render a Markdown document in raw mode
}
