% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_meta_get}
\alias{gh_meta_get}
\title{This gives some information about GitHub.com, the service.}
\usage{
gh_meta_get(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
This gives some information about GitHub.com, the service.
}
