% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_notifications_get}
\alias{gh_notifications_get}
\title{List your notifications.}
\usage{
gh_notifications_get(...)
}
\arguments{
\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
List all notifications for the current user, grouped by repository.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
