% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_notifications_get_threads_subscription}
\alias{gh_notifications_get_threads_subscription}
\title{Get a Thread Subscription.}
\usage{
gh_notifications_get_threads_subscription(id, ...)
}
\arguments{
\item{id}{Id of thread.}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Get a Thread Subscription.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
