% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_notifications_patch_threads_by_id}
\alias{gh_notifications_patch_threads_by_id}
\title{Mark a thread as read}
\usage{
gh_notifications_patch_threads_by_id(id, ...)
}
\arguments{
\item{id}{Id of thread.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Mark a thread as read
}
