% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_orgs_get_members}
\alias{gh_orgs_get_members}
\title{Members list.}
\usage{
gh_orgs_get_members(org, ...)
}
\arguments{
\item{org}{Name of organisation.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List all users who are members of an organization. A member is a user tha
belongs to at least 1 team in the organization. If the authenticated user is
also an owner of this organization then both concealed and public members will
be returned. If the requester is not an owner of the organization the query
will be redirected to the public members list.
}
