% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_orgs_patch_by_org}
\alias{gh_orgs_patch_by_org}
\title{Edit an Organization.}
\usage{
gh_orgs_patch_by_org(org, body, ...)
}
\arguments{
\item{org}{Name of organisation.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Edit an Organization.
}
