% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_rate_limit_get}
\alias{gh_rate_limit_get}
\title{Get your current rate limit status}
\usage{
gh_rate_limit_get(...)
}
\arguments{
\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Note: Accessing this endpoint does not count against your rate limit.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
