% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_delete_git_refs_by_ref}
\alias{gh_repos_delete_git_refs_by_ref}
\title{Delete a Reference}
\usage{
gh_repos_delete_git_refs_by_ref(owner, repo, ref, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{ref}{ref}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Example: Deleting a branch: DELETE
/repos/octocat/Hello-World/git/refs/heads/feature-a Example: Deleting a tag:
DELETE /repos/octocat/Hello-World/git/refs/tags/v1.0
}
