% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_compare_by_base_id_head_id}
\alias{gh_repos_get_compare_by_base_id_head_id}
\title{Compare two commits}
\usage{
gh_repos_get_compare_by_base_id_head_id(owner, repo, baseId, headId, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{baseId}{baseId}

\item{headId}{headId}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Compare two commits
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
