% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_contents_by_path}
\alias{gh_repos_get_contents_by_path}
\title{Get contents.}
\usage{
gh_repos_get_contents_by_path(owner, repo, path, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{path}{path}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
This method returns the contents of a file or directory in a repository. Files
and symlinks support a custom media type for getting the raw content.
Directories and submodules do not support custom media types. Note: This API
supports files up to 1 megabyte in size. Here can be many outcomes. For details
see "http://developer.github.com/v3/repos/contents/"
}
