% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_deployments_statuses}
\alias{gh_repos_get_deployments_statuses}
\title{Users with pull access can view deployment statuses for a deployment}
\usage{
gh_repos_get_deployments_statuses(owner, repo, id, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{id}{The Deployment ID to list the statuses from.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Users with pull access can view deployment statuses for a deployment
}
