% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_languages}
\alias{gh_repos_get_languages}
\title{List languages.}
\usage{
gh_repos_get_languages(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List languages for the specified repository. The value on the right of a
language is the number of bytes of code written in that language.
}
