% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_pulls_files}
\alias{gh_repos_get_pulls_files}
\title{List pull requests files.}
\usage{
gh_repos_get_pulls_files(owner, repo, number, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{number}{Id of pull.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List pull requests files.
}
