% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_releases}
\alias{gh_repos_get_releases}
\title{Users with push access to the repository will receive all releases (i.e., published releases and draft releases). Users with pull access will receive published releases only}
\usage{
gh_repos_get_releases(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Users with push access to the repository will receive all releases (i.e., published releases and draft releases). Users with pull access will receive published releases only
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
