% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_get_stats_code_frequency}
\alias{gh_repos_get_stats_code_frequency}
\title{Get the number of additions and deletions per week.}
\usage{
gh_repos_get_stats_code_frequency(owner, repo, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Returns a weekly aggregate of the number of additions and deletions pushed to a
repository.
}
