% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_patch_git_refs_by_ref}
\alias{gh_repos_patch_git_refs_by_ref}
\title{Update a Reference}
\usage{
gh_repos_patch_git_refs_by_ref(owner, repo, ref, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{ref}{ref}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Update a Reference
}
