% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_patch_pulls_by_number}
\alias{gh_repos_patch_pulls_by_number}
\title{Update a pull request.}
\usage{
gh_repos_patch_pulls_by_number(owner, repo, number, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{number}{Id of pull.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Update a pull request.
}
