% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_patch_releases_by_id}
\alias{gh_repos_patch_releases_by_id}
\title{Users with push access to the repository can edit a release}
\usage{
gh_repos_patch_releases_by_id(owner, repo, id, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{id}{id}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Users with push access to the repository can edit a release
}
