% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_post_deployments}
\alias{gh_repos_post_deployments}
\title{Users with push access can create a deployment for a given ref}
\usage{
gh_repos_post_deployments(owner, repo, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Users with push access can create a deployment for a given ref
}
