% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_post_git_commits}
\alias{gh_repos_post_git_commits}
\title{Create a Commit.}
\usage{
gh_repos_post_git_commits(owner, repo, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Create a Commit.
}
