% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_put_notifications}
\alias{gh_repos_put_notifications}
\title{Mark notifications as read in a repository.}
\usage{
gh_repos_put_notifications(owner, repo, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Marking all notifications in a repository as "read" removes them from the
default view on GitHub.com.
}
