% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_repos_put_pulls_merge}
\alias{gh_repos_put_pulls_merge}
\title{Merge a pull request (Merge Button's)}
\usage{
gh_repos_put_pulls_merge(owner, repo, number, body, ...)
}
\arguments{
\item{owner}{Name of repository owner.}

\item{repo}{Name of repository.}

\item{number}{Id of pull.}

\item{body}{body}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Merge a pull request (Merge Button's)
}
