% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_search_get_issues}
\alias{gh_search_get_issues}
\title{Find issues by state and keyword. (This method returns up to 100 results per page.)}
\usage{
gh_search_get_issues(q, ...)
}
\arguments{
\item{q}{The q search term can also contain any combination of the supported issue
search qualifiers:}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Find issues by state and keyword. (This method returns up to 100 results per page.)
}
