% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_teams_get_members}
\alias{gh_teams_get_members}
\title{List team members.}
\usage{
gh_teams_get_members(teamId, ...)
}
\arguments{
\item{teamId}{Id of team.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
In order to list members in a team, the authenticated user must be a member of
the team.
}
