% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_delete_keys_by_key_id}
\alias{gh_user_delete_keys_by_key_id}
\title{Delete a public key. Removes a public key. Requires that you are authenticated via Basic Auth or via OAuth with at least admin:public_key scope.}
\usage{
gh_user_delete_keys_by_key_id(keyId, ...)
}
\arguments{
\item{keyId}{ID of key.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Delete a public key. Removes a public key. Requires that you are authenticated via Basic Auth or via OAuth with at least admin:public_key scope.
}
