% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_get_following}
\alias{gh_user_get_following}
\title{List who the authenticated user is following.}
\usage{
gh_user_get_following(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List who the authenticated user is following.
}
