% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_get_following_by_username}
\alias{gh_user_get_following_by_username}
\title{Check if you are following a user.}
\usage{
gh_user_get_following_by_username(username, ...)
}
\arguments{
\item{username}{Name of user.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Check if you are following a user.
}
