% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_get_keys_by_key_id}
\alias{gh_user_get_keys_by_key_id}
\title{Get a single public key.}
\usage{
gh_user_get_keys_by_key_id(keyId, ...)
}
\arguments{
\item{keyId}{ID of key.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Get a single public key.
}
