% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_get_subscriptions}
\alias{gh_user_get_subscriptions}
\title{List repositories being watched by the authenticated user.}
\usage{
gh_user_get_subscriptions(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List repositories being watched by the authenticated user.
}
