% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_get_subscriptions_by_repo}
\alias{gh_user_get_subscriptions_by_repo}
\title{Check if you are watching a repository.}
\usage{
gh_user_get_subscriptions_by_repo(owner, repo, ...)
}
\arguments{
\item{owner}{Name of the owner.}

\item{repo}{Name of repository.}

\item{...}{Additional parameters to pass to \code{\link[gh]{gh}}. See details.}
}
\description{
Check if you are watching a repository.
}
\details{
Additional parameters that can be passed:
\describe{
\item{type}{ It takes values in: all, public, private, forks, sources, member.
The default is: all}
\item{Accept}{Is used to set specified media type.  }
}
}
