% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_user_get_teams}
\alias{gh_user_get_teams}
\title{List all of the teams across all of the organizations to which the authenticated user belongs. This method requires user or repo scope when authenticating via OAuth.}
\usage{
gh_user_get_teams(...)
}
\arguments{
\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
List all of the teams across all of the organizations to which the authenticated user belongs. This method requires user or repo scope when authenticating via OAuth.
}
