% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh3.R
\name{gh_users_get_following_by_target_user}
\alias{gh_users_get_following_by_target_user}
\title{Check if one user follows another.}
\usage{
gh_users_get_following_by_target_user(username, targetUser, ...)
}
\arguments{
\item{username}{Name of user.}

\item{targetUser}{Name of user.}

\item{...}{Additional parameters to pass to gh::gh.}
}
\description{
Check if one user follows another.
}
